#11: fixed issue raised by comments on PR, added comments to a method
